PR 리뷰 반영 - 클래스분리 https://github.com/next-step/kotlin-lotto/pull/728#discussion_r1235445044
